Seperates myItem and item functionalities
