+++ frontEnd/src/myPosts.jsx
@@ -1,6 +1,6 @@
 import { useState } from "react";
 import "./App.css";
-import Item from "./components/item";
+import MyItem from "./components/myItem.jsx";
 import Navigation from "./components/nav";
 import { Context } from "./App";
 import { useContext } from "react";
@@ -41,7 +41,7 @@ function MyPosts() {
         ) : isDonationList ? (
           donations?.map((item) => {
             return (
-              <Item
+              <MyItem
                 onPostChange={() => {
                   fetchMyPosts();
                 }}
@@ -56,7 +56,7 @@ function MyPosts() {
         ) : (
           requests?.map((item) => {
             return (
-              <Item
+              <MyItem
                 onPostChange={() => {
                   fetchMyPosts();
                 }}
